[coin.move] CoinMetadata getter functions (#6911)
- Currently, `CoinMetadata` can be accessed for off-chain consumption,
however there is a class of use-cases where we would like to consume the
metadata fields (decimals, name, symbol, etc.) on-chain. For example, a
protocol might want to perform checks on the metadata corresponding to a
`CoinType` on-chain, e.g. check that `decimals==8`. Currently this is
impossible because the metadata fields are opaque. The following getters
allow inspection of `CoinMetadata` fields.