Bump fastcrypto version (#8228)
This PR bumps Sui to the most recent version of fastcrypto. There's been
a few changes in the traits in fastcrypto, namely that
`signature::Signer` and `signature::Signature` are no longer used. In
fastcrypto, these were replaced with new internal traits `Signer` and
`ToFromBytes` respectively.

The new `Signer` trait does not have a `try_sign` function like the
previous one, but the only place this ever returns an error is
`SuiInnerSiganture::new` if a serialised signature does not have the
correct size. This should never happen, so we can safely remove this
error.

Besides the removed error, this does not introduce any functional
changes.