Swarm config refactoring (#12117)
This PR does a series of refactoring on how we construct node config and
swarm configs to make it cleaner:
1. Renames utils.rs in sui-config/ to local_ip_utils.rs, and moved all
the IP and port management port into it. Now it contains a manager to
generate new IPs and new ports for simtest; for non-simtest it always
use localhost and new available port.
2. Previously we start fullnode inside TestCluster, which runs on the
client node in simtest. This created a lot of tech debt. This PR moves
all fullnode creation into Swarm. Now we can start arbitrary number of
fullnodes in a Swarm. When starting TestCluster, it uses Swarm to launch
a fullnode, and keep a copy of the fullnode handle locally for rpc
access.
3. Also added the ability to start more fullnodes after a test cluster
started running, inside swarm. This means we could also run those
fullnodes in a container, managed by the swarm.
4. Heavy use of the builder pattern: added ValidatorGenesisConfigBuilder
to build a ValidatorGenesisConfig; added ValidatorConfigBuilder and
FullnodeConfigBuilder to build their node configs.
5. Removed a number of unused fields in ConfigBuilder and SwarmBuilder.

This this PR, we should be able to start removing all the code that
spawns individual nodes instead of inside a container.