[framework] don't emit event on purchase with PurchaseCap (#12821)
## Motivation

PurchaseCap functionality is created specifically for extensions and
logic which may or may not be considered trading. And emitting an event
when "purchasing" with the Cap breaks indexers and makes little sense -
ItemListed is not emitted, ItemDelisted as well.

Currently, all extensions rely on custom events for this purpose.

## Description 

Changes the body of the `kiosk::purchase_with_cap` to NOT emit the
ItemPurchased event. Everything else is left intact.

## Test Plan 

Current tests pass.

### Type of Change (Check all that apply)

- [ ] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

- The `kiosk::purchase_with_cap` method for [Sui
Kiosk](https://docs.sui.io/build/sui-kiosk) no longer emits the
`ItemPurchased` event. This is a network optimization update that should
have minimal impact. If the logic of your application does rely on this
event, you must change it to leverage custom events instead.