[move] Handling recently introduced warnings in the framework code (#13866)
## Description 

Recently introduced warnings caused a lot of noise when building
framework code itself. While only meaningful to a rather small number of
devs (working on the framework code), it was nevertheless annoying and
this PR is a cleanup attempt. When compiling user code, these warnings
are suppressed anyway - hence no user impact.

The strategy was to use suppression unless (and ALWAYS in the governance
code) unless it seemed safe to remove a given construct - in particular,
if there was an unused constant (particularly indicating an error) and
there are no native functions in the module.

## Test Plan 

All tests in sui-framework-tests must pass

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes