Add zklogin check and verify functions (#13817)
## Description 

Add functions which allows a user to prove that a certain identity was
used to create a zklogin address and link this identity to the address
(VerifiedID).

## Test Plan 

Unit tests

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [x] protocol change
- [x] user-visible impact
- [ ] breaking change for a client SDKs
- [x] breaking change for FNs (FN binary must upgrade)
- [x] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Add a `verify_zklogin_id` function to the SUI framework which allows
users to prove that a certain identity was used to create a zklogin
address and link this identity to their address. The function returns a
verified identity object which maybe used to authenticate towards smart
contracts. A similar function where the user only reveals who issued the
identity used to create a zklogin address, `verify_zklogin_iss` is also
added.

---------

Co-authored-by: Sam Blackshear <sam.blackshear@gmail.com>
Co-authored-by: Todd Nowacki <tmn@mystenlabs.com>