[transfer-to-object] Update `receive` to follow private generic rules (#14723)
## Description 

Updates the ability to receive objects so that this follows similar
rules to `transfer` and `public_transfer`:
* `receive` can receive `key`-only objects, but this function can only
be called in the module that defines the type being received.
* `public_receive` can receive `key + store` objects, and can be called
outside of the module that defines the type that is being received.

Note that I still need to add protocol versioning for this change since
we need to version:
* The sui-verifier checks; and 
* The checks for calls from a PTB into `receive`
However this will necessitate cutting a new execution version (which we
plan on doing anyways).

## Test Plan 

Added a number of new tests checking that we properly enforce private
generics.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [X] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes

Changes the behavior of `sui::transfer::receive` to obey the same custom
rules around `key`-only objects as `sui::transfer::transfer` and adds a
`sui::transfer::public_receive` function for `key + store` objects.